+++ client/src/main/java/replicant/messages/UnsubscribeMessage.java
@@ -1,0 +1,26 @@
+package replicant.messages;
+
+import javax.annotation.Nonnull;
+import jsinterop.annotations.JsOverlay;
+import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsType;
+
+@JsType( isNative = true, namespace = JsPackage.GLOBAL, name = "Object" )
+@SuppressWarnings( { "NullableProblems", "unused" } )
+public final class UnsubscribeMessage
+  extends ClientToServerMessage
+{
+  @Nonnull
+  private String ch;
+
+  @JsOverlay
+  @Nonnull
+  public static UnsubscribeMessage create( final int req, @Nonnull final String ch )
+  {
+    final UnsubscribeMessage message = new UnsubscribeMessage();
+    message.type = "unsub";
+    message.req = req;
+    message.ch = ch;
+    return message;
+  }
+}